Move tooltip render into Editor component
